requirements.txt
python-dateutil==2.4.2
gspread==0.2.3
pytz==2014.10
PyYAML==3.11
requests==2.5.3